Move replayerInfo type and functions to separate file
